+++ src/libigl/igl/copyleft/cgal/subdivide_segments.cpp
@@ -1,9 +1,9 @@
 // This file is part of libigl, a simple c++ geometry processing library.
-// 
+//
 // Copyright (C) 2016 Alec Jacobson <alecjacobson@gmail.com>
-// 
-// This Source Code Form is subject to the terms of the Mozilla Public License 
-// v. 2.0. If a copy of the MPL was not distributed with this file, You can 
+//
+// This Source Code Form is subject to the terms of the Mozilla Public License
+// v. 2.0. If a copy of the MPL was not distributed with this file, You can
 // obtain one at http://mozilla.org/MPL/2.0/.
 #include "subdivide_segments.h"
 #include "row_to_point.h"
@@ -17,10 +17,10 @@
 #include <vector>
 
 template <
-  typename DerivedV, 
+  typename DerivedV,
   typename DerivedE,
-  typename Kernel, 
-  typename DerivedVI, 
+  typename Kernel,
+  typename DerivedVI,
   typename DerivedEI,
   typename DerivedJ,
   typename DerivedIM>
@@ -83,7 +83,7 @@ IGL_INLINE void igl::copyleft::cgal::subdivide_segments(
     }
     // remove duplicates
     steiner[i].erase(
-      std::unique(steiner[i].begin(), steiner[i].end()), 
+      std::unique(steiner[i].begin(), steiner[i].end()),
       steiner[i].end());
     {
       int s = E(i,0);
@@ -110,7 +110,7 @@ IGL_INLINE void igl::copyleft::cgal::subdivide_segments(
       assign_scalar(V(i,j),VI(i,j));
     }
   }
-  for(int i = 0;i<S.size();i++)
+  for( size_t i = 0;i<S.size();i++)
   {
     assign_scalar(S[i].x(),VI(n+i,0));
     assign_scalar(S[i].y(),VI(n+i,1));
